--- frontend-svelte/src/components/Dummy.svelte
@@ -1,9 +1,0 @@
-<script>
-    export let value, setValue, theme;
-    export let choices;
-</script>
-
-<div class="dummy" {theme}>DUMMY</div>
-
-<style lang="postcss">
-</style>
