--- frontend-svelte/src/components/output/Audio.svelte
@@ -1,10 +1,0 @@
-<script>
-  export let value, theme;
-</script>
-
-<audio {theme} controls>
-  <source src={value} />
-</audio>
-
-<style lang="postcss">
-</style>
