--- frontend-svelte/src/components/output/Html.svelte
@@ -1,10 +1,0 @@
-<script>
-  export let value, theme;
-</script>
-
-<div
-  class="output-html"
-  {theme}
->
-  {@html value}
-</div>
