--- frontend-svelte/src/components/output/Image.svelte
@@ -1,10 +1,0 @@
-<script>
-  export let value, theme;
-</script>
-
-<div class="output-image w-full h-60 flex justify-center items-center bg-gray-200 dark:bg-gray-600 relative" {theme}>
-  <img class="w-full h-full object-contain" src={value} />
-</div>
-
-<style lang="postcss">
-</style>
