+++ frontend/src/components/output/Label/Component.svelte
@@ -25,8 +25,8 @@
       <div class="confidences flex flex-grow flex-col items-baseline">
         {#each value.confidences as confidence_set, i}
           <div
-            class="confidence flex justify-end overflow-hidden whitespace-nowrap h-7 mb-2 px-1"
-            style={"min-width: calc(" +
+          class="confidence flex justify-end items-center overflow-hidden whitespace-nowrap h-7 mb-2 px-1"
+          style={"min-width: calc(" +
               Math.round(confidence_set.confidence * 100) +
               "% - 12px)"}
             key={i}
@@ -42,10 +42,10 @@
 <style lang="postcss">
   .output-label[theme="default"] {
     .label {
-      @apply text-sm h-5;
+      @apply text-base h-7;
     }
     .confidence {
-      @apply font-mono box-border border-b-2 border-gray-300 bg-gray-200 dark:bg-gray-500 dark:border-gray-600 text-sm h-5 font-semibold rounded;
+      @apply font-mono box-border border-b-2 border-gray-300 bg-gray-200 dark:bg-gray-500 dark:border-gray-600 text-sm h-7 font-semibold rounded;
     }
     .confidence:first-child {
       @apply border-yellow-600 bg-yellow-500 dark:bg-red-600 border-red-700 text-white;
